🚚 rename: restructure page routes for admin separation
