+++ src/app/(main)/(support)/notice/[id]/page.tsx
