+++ src/app/(main)/(support)/terms/page.tsx
