+++ src/app/(main)/character/page.tsx
