+++ src/app/(main)/community/[id]/page.tsx
