+++ src/app/(main)/community/update/[id]/page.tsx
