+++ src/app/(main)/community/write/page.tsx
