+++ src/app/(main)/favicon.ico
