+++ src/app/(main)/layout.tsx
@@ -7,7 +7,7 @@ import AdBannerModal from '@/components/features/user/alarm/AdBannerModal';
 import { ToastProvider } from '@/components/common/ToastProvider';
 
 const pretendard = localFont({
-  src: '../../public/fonts/PretendardVariable.woff2',
+  src: '../../../public/fonts/PretendardVariable.woff2',
   display: 'swap',
   weight: '45 920',
   variable: '--font-pretendard',
