+++ src/app/(main)/live/LiveContent.tsx
