+++ src/app/(main)/user/[id]/page.tsx
